requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
aiogram==3.15.0
httpx==0.28.1
pydantic-settings==2.12.0
python-dotenv==1.2.1
bcrypt==5.0.0
tortoise-orm==0.25.3
aerich[cli]==0.9.2
aiosqlite==0.22.1
asyncpg==0.30.0
pytest==9.0.2
pytest-asyncio==1.3.0
tomlkit
jinja2==3.1.5
python-multipart==0.0.9
aiofiles==24.1.0